Added a endpoint to retrieve all loan requests
Implemented a new API endpoint /api/get_all_loan_request that connects to MongoDB and retrieves all loan requests using the Mongoose Loan model. If successful, it responds with a JSON array containing the retrieved data. In case of an error, it returns an appropriate error message along with a 500 status code.

This endpoint allows the client to fetch and display all loan requests for further processing and analysis.