Add route to get repayment schedule by transaction_ID
Added a new route in the API to retrieve repayment schedule details using the provided transaction ID. The route connects to MongoDB and fetches the data associated with the given transaction ID from the Loan collection. In case of success, the fetched data is sent as a JSON response, and in case of an error, an appropriate error message is returned along with a status code.